added missing double casting to Poisson lambda in ngdepth

+++ src/net/derkholm/nmica/extra/app/seq/nextgen/DepthMovingAverage.java
@@ -47,7 +47,7 @@ public class DepthMovingAverage extends SAMProcessor {
 	public void setRefLengths(File f) throws BioException, IOException {
 		super.setRefLengths(f);
 		for (String name : this.refSeqLengths.keySet()) {
-			nullDistributions.put(name, new Poisson(this.refSeqLengths.get(name) / this.windowSize, randomEngine));
+			nullDistributions.put(name, new Poisson((double)this.refSeqLengths.get(name) / (double)this.windowSize, randomEngine));
 		}
 	}
